gateway/requirements.txt
paho-mqtt>=1.6
mysql-connector-python>=8.0
